move formatSize to format file
