+++ .changeset/orange-geese-attend.md
@@ -1,0 +1,5 @@
+---
+"claude-dev": patch
+---
+
+cleanup
