+++ webview-ui/src/components/chat/TaskHeader.tsx
@@ -5,7 +5,7 @@ import { mentionRegexGlobal } from "../../../../src/shared/context-mentions"
 import { ClineMessage } from "../../../../src/shared/ExtensionMessage"
 import { useExtensionState } from "../../context/ExtensionStateContext"
 import { formatLargeNumber } from "../../utils/format"
-import { formatSize } from "../../utils/size"
+import { formatSize } from "../../utils/format"
 import { vscode } from "../../utils/vscode"
 import Thumbnails from "../common/Thumbnails"
 import { normalizeApiConfiguration } from "../settings/ApiOptions"
