+++ webview-ui/src/components/history/HistoryView.tsx
@@ -5,7 +5,7 @@ import { Virtuoso } from "react-virtuoso"
 import { memo, useMemo, useState, useEffect, useCallback } from "react"
 import Fuse, { FuseResult } from "fuse.js"
 import { formatLargeNumber } from "../../utils/format"
-import { formatSize } from "../../utils/size"
+import { formatSize } from "../../utils/format"
 import { ExtensionMessage } from "../../../../src/shared/ExtensionMessage"
 import { useEvent } from "react-use"
 import DangerButton from "../common/DangerButton"
