+++ webview-ui/src/utils/format.ts
@@ -1,3 +1,5 @@
+import prettyBytes from "pretty-bytes"
+
 export function formatLargeNumber(num: number): string {
 	if (num >= 1e9) {
 		return (num / 1e9).toFixed(1) + "b"
@@ -33,4 +35,12 @@ export function formatTimestamp(timestamp: string): string {
 	})
 
 	return dateFormatter.format(date)
+}
+
+export function formatSize(bytes?: number) {
+	if (bytes === undefined) {
+		return "--kb"
+	}
+
+	return prettyBytes(bytes)
 }
