--- webview-ui/src/utils/size.ts
@@ -1,9 +1,0 @@
-import prettyBytes from "pretty-bytes"
-
-export function formatSize(bytes?: number) {
-	if (bytes === undefined) {
-		return "--kb"
-	}
-
-	return prettyBytes(bytes)
-}
